# fixed dense times diagonal matrix multiplication