Fixing #27176 to make dense times diagonal matrix multiplication return dense matrices.

Some basic timings:

Before:
```
julia> A = Tridiagonal(rand(2000,2000))
julia> B = rand(2000,2000)
julia> @time A*B
  0.075236 seconds (6 allocations: 30.518 MiB, 87.87% gc time)
julia> @time B*A
  4.515886 seconds (50 allocations: 66.017 MiB, 0.07% gc time)
```
After:
```
julia> A = Tridiagonal(rand(2000,2000))
julia> B = rand(2000,2000)
julia> @time A*B
  0.023720 seconds (6 allocations: 30.518 MiB, 7.70% gc time)
julia> @time B*A
  0.017339 seconds (6 allocations: 30.518 MiB, 10.38% gc time)
```

This also affects dense times Bidiagonal and SymTriDiagonal matrix multiplication.

----------
```
julia> versioninfo()
Julia Version 0.7.0-beta2.0
Commit b145832402* (2018-07-13 19:54 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: Intel(R) Core(TM) i7-6600U CPU @ 2.60GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.0 (ORCJIT, skylake)
```