@ViralBShah 2019-07-06
Bump. Would be nice to get this in. Performance of the reported case is still poor.